chore: Fixed productVisibilitySet export
